Allow delayed gift cards to be correctly sent
Gift cards need to be able to be sent in a delayed manner. To that end,
introduce a two stage "ready_for_release" => "deliver" flow.

Once an order is paid for, a gift card is marked as "ready for release"
-- without this flag set (and it defaults to false in the database)
it'll raise an error if you attempt to send it.

This flag is set by the payment_decorator following the same logic as
previously once the payment goes to the "Completed" state.

Add a rake task (intended to be used as a cron job) which then
calls `deliver!` on those gift cards that have not previously been sent
that have a date to send less than time.now.